==> Part3_generate_syth_stock_from_model.ipynb <==
import os
# Set & view the working directory to ensure runtime is executng in expected directory
cwd = os.getcwd()
print(f"Current working directory: {cwd}")
os.chdir("C:/Users/angusf/source/repos/synth_nonin") #note the use of forward slashes

from sdv.sequential import PARSynthesizer
synthesizer = PARSynthesizer.load('IntraDayStockSynthesiser.pkl') # Load the model we trained  in step 2

synthetic_data = synthesizer.sample(num_sequences=12, sequence_length=300) #generate 12 sequences (of Stock Symbols) of 96 rows each (8 hours worth of 5 minute data samples)
synthetic_data.head()

#sort the dataset so we can better see the sequences (which in my experience are not always correct for time series)
synthetic_data.sort_values(['Symbol', 'Combineddatetime'])

synthetic_data.to_csv('synthetic_Intraday PARstock.csv', index=False) #save the synthetic data to a csv file
synthetic_data['Symbol'].unique() #list the synthetic stock symbols
