==> Part3_1_plot_synth_data.ipynb <==
import pandas

synthetic_stock_data = pandas.read_csv('./synthetic_Intraday_PARstock.csv')
synthetic_stock_data['Combineddatetime']=pandas.to_datetime(synthetic_stock_data['Combineddatetime']) # Convert to datetime for correct future usage in plotting etc
synthetic_stock_data['Time']=pandas.to_datetime(synthetic_stock_data['Time'])


import matplotlib.pyplot as plt

# plot the synth data grouped by stock symbol
grouped_data = synthetic_stock_data.groupby('Symbol')

# plot the data for each stock symbol
for symbol, data in grouped_data:
    data.plot(x='Combineddatetime', y='Open', label="Synthesised data for synthetic symbol:" + symbol)

# add a legend to the plot

plt.legend()
plt.show()
